chore: move utils into unit test folder
