+++ __tests__/unit/specs/components/Footer.spec.js
@@ -1,7 +1,7 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils'
 import sinon from 'sinon'
 import AppFooter from '@/components/AppFooter'
-import { useI18n } from '../../../__utils__/i18n'
+import { useI18n } from '../../__utils__/i18n'
 
 global.GIT_VERSION = '43496685190e3e768c3f5b1bc322ff8b7ed4c696'
 global.GIT_DATE = '2018-01-01'
