+++ __tests__/unit/specs/components/header/Languages.spec.js
@@ -1,7 +1,7 @@
 import { mount, createLocalVue } from '@vue/test-utils'
 import mixins from '@/mixins'
 import { HeaderLanguagesDesktop } from '@/components/header/languages'
-import { useI18n } from '../../../../__utils__/i18n'
+import { useI18n } from '../../../__utils__/i18n'
 import Vuex from 'vuex'
 import moment from 'moment'
