+++ __tests__/unit/specs/components/header/ToggleChart.spec.js
@@ -2,7 +2,7 @@ import { mount, createLocalVue } from '@vue/test-utils'
 import mixins from '@/mixins'
 
 import ToggleChart from '@/components/header/toggles/ToggleChart'
-import { useI18n } from '../../../../__utils__/i18n'
+import { useI18n } from '../../../__utils__/i18n'
 import Vuex from 'vuex'
 
 describe('Components > Header > ToggleChart', () => {
