+++ __tests__/unit/specs/components/wallet/Delegate.spec.js
@@ -2,7 +2,7 @@ import { mount, createLocalVue, RouterLinkStub } from '@vue/test-utils'
 import mixins from '@/mixins'
 
 import WalletDelegate from '@/components/wallet/Delegate'
-import { useI18n } from '../../../../__utils__/i18n'
+import { useI18n } from '../../../__utils__/i18n'
 import Vuex from 'vuex'
 
 describe('Components > Wallet > Delegate', () => {
